pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
# momex pin: v0.2.2

[project]
name = "koa"
version = "0.1.1"
description = "A personal AI agent framework with state-driven agents and ReAct orchestration"
readme = "README.md"
license = {text = "BSL-1.1"}
requires-python = ">=3.12"
authors = [
    { name = "Koa Contributors" }
]
keywords = [
    "agent",
    "llm",
    "ai",
    "workflow",
    "orchestration",
    "state-machine",
    "multi-agent",
    "framework",
    "zero-code",
    "yaml",
    "openai",
    "anthropic",
    "claude",
    "gpt",
    "streaming"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]

# Core dependencies
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "asyncpg>=0.27",
    "httpx>=0.25",
    "python-dateutil>=2.8",
    "fastapi>=0.100",
    "uvicorn>=0.20",
    "momex @ git+https://github.com/xiaoyu-work/momex.git@v0.2.2",
    "jinja2>=3.1",
    "litellm>=1.40",
    "trafilatura>=2.0",
    "alembic>=1.12",
    "sqlalchemy>=2.0",
    "psycopg2-binary>=2.9",
    "supabase>=2.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0"]
anthropic = ["anthropic>=0.18"]
gemini = ["google-generativeai>=0.3"]
mcp = ["mcp>=1.2"]
all = [
    "openai>=1.0",
    "anthropic>=0.18",
    "google-generativeai>=0.3",
    "mcp>=1.2",
]

[project.scripts]
koa = "koa.server:main"

[project.urls]
Homepage = "https://github.com/withkoi/koa"
Documentation = "https://withkoi.github.io/koa"
Repository = "https://github.com/withkoi/koa"
Issues = "https://github.com/withkoi/koa/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["koa"]

[tool.black]
line-length = 100
target-version = ["py312", "py313"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
disable_error_code = [
    "attr-defined",
    "no-any-return",
    "assignment",
    "union-attr",
    "import-untyped",
    "annotation-unchecked",
    "arg-type",
    "return-value",
    "call-arg",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "integration: end-to-end tests using a real LLM (requires INTEGRATION_TEST_API_KEY)",
    "communication: agents in the communication domain (email, slack, discord, twitter, linkedin)",
    "productivity: agents in the productivity domain (calendar, todo, briefing, notion, google_workspace, cloud_storage, github, cron)",
    "lifestyle: agents in the lifestyle domain (expense, smarthome, shipping, spotify, youtube, image)",
    "travel: agents in the travel domain (trip_planner, maps)",
]
